pyproject.toml
[project]
name = "agentic-sales-copilot"
version = "0.1.0"
description = "An autonomous, multi-agent system designed to automate sales email outreach, intelligently process replies, and provide a human-in-the-loop interface for sales representatives via Slack."
authors = [
    {name = "Harshad Lohande",email = "lohande.harshad@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai (>=1.98.0,<2.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "openai-agents (>=0.2.4,<0.3.0)",
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "sendgrid (>=6.12.4,<7.0.0)",
    "slack-sdk (>=3.36.0,<4.0.0)",
    "aiohttp (>=3.12.15,<4.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "python-json-logger (>=3.3.0,<4.0.0)"
]

[tool.poetry]
package-mode = false

[tool.poetry.scripts]
start = "app.main:start"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = ["."]
